Extract TypeHandlingJdbcConfig from BaseJdbcConfig
