+++ presto-base-jdbc/src/test/java/io/prestosql/plugin/jdbc/TestBaseJdbcConfig.java
@@ -36,8 +36,7 @@ public class TestBaseJdbcConfig
                 .setConnectionUrl(null)
                 .setCaseInsensitiveNameMatching(false)
                 .setCaseInsensitiveNameMatchingCacheTtl(new Duration(1, MINUTES))
-                .setJdbcTypesMappedToVarchar(null)
-                .setUnsupportedTypeHandling(UnsupportedTypeHandling.IGNORE)
+                .setJdbcTypesMappedToVarchar("")
                 .setMetadataCacheTtl(Duration.valueOf("0m"))
                 .setCacheMissing(false));
     }
@@ -50,7 +49,6 @@ public class TestBaseJdbcConfig
                 .put("case-insensitive-name-matching", "true")
                 .put("case-insensitive-name-matching.cache-ttl", "1s")
                 .put("jdbc-types-mapped-to-varchar", "mytype,struct_type1")
-                .put("unsupported-type-handling", "CONVERT_TO_VARCHAR")
                 .put("metadata.cache-ttl", "1s")
                 .put("metadata.cache-missing", "true")
                 .build();
@@ -60,7 +58,6 @@ public class TestBaseJdbcConfig
                 .setCaseInsensitiveNameMatching(true)
                 .setCaseInsensitiveNameMatchingCacheTtl(new Duration(1, SECONDS))
                 .setJdbcTypesMappedToVarchar("mytype, struct_type1")
-                .setUnsupportedTypeHandling(UnsupportedTypeHandling.CONVERT_TO_VARCHAR)
                 .setMetadataCacheTtl(Duration.valueOf("1s"))
                 .setCacheMissing(true);
