+++ presto-base-jdbc/src/test/java/io/prestosql/plugin/jdbc/TestingH2JdbcModule.java
@@ -34,6 +34,7 @@ class TestingH2JdbcModule
     public void configure(Binder binder)
     {
         configBinder(binder).bindConfig(BaseJdbcConfig.class);
+        configBinder(binder).bindConfig(TypeHandlingJdbcConfig.class);
     }
 
     @Provides
