+++ presto-memsql/src/main/java/io/prestosql/plugin/memsql/MemSqlClientModule.java
@@ -24,6 +24,7 @@ import io.prestosql.plugin.jdbc.ConnectionFactory;
 import io.prestosql.plugin.jdbc.DriverConnectionFactory;
 import io.prestosql.plugin.jdbc.ForBaseJdbc;
 import io.prestosql.plugin.jdbc.JdbcClient;
+import io.prestosql.plugin.jdbc.TypeHandlingJdbcConfig;
 import io.prestosql.plugin.jdbc.credential.CredentialProvider;
 import org.mariadb.jdbc.Driver;
 
@@ -43,6 +44,7 @@ public class MemSqlClientModule
         binder.bind(Key.get(JdbcClient.class, ForBaseJdbc.class))
                 .to(MemSqlClient.class).in(Scopes.SINGLETON);
         ensureCatalogIsEmpty(buildConfigObject(BaseJdbcConfig.class).getConnectionUrl());
+        configBinder(binder).bindConfig(TypeHandlingJdbcConfig.class);
         configBinder(binder).bindConfig(MemSqlConfig.class);
     }
