+++ presto-postgresql/src/main/java/io/prestosql/plugin/postgresql/PostgreSqlClientModule.java
@@ -28,6 +28,7 @@ import io.prestosql.plugin.jdbc.DriverConnectionFactory;
 import io.prestosql.plugin.jdbc.ForBaseJdbc;
 import io.prestosql.plugin.jdbc.JdbcClient;
 import io.prestosql.plugin.jdbc.SessionPropertiesProvider;
+import io.prestosql.plugin.jdbc.TypeHandlingJdbcConfig;
 import io.prestosql.plugin.jdbc.credential.CredentialProvider;
 import org.postgresql.Driver;
 
@@ -43,6 +44,7 @@ public class PostgreSqlClientModule
         binder.bind(Key.get(JdbcClient.class, ForBaseJdbc.class))
                 .to(PostgreSqlClient.class).in(Scopes.SINGLETON);
         configBinder(binder).bindConfig(BaseJdbcConfig.class);
+        configBinder(binder).bindConfig(TypeHandlingJdbcConfig.class);
         configBinder(binder).bindConfig(PostgreSqlConfig.class);
         configBinder(binder).bindConfig(DecimalConfig.class);
