+++ presto-sqlserver/src/main/java/io/prestosql/plugin/sqlserver/SqlServerClientModule.java
@@ -25,6 +25,7 @@ import io.prestosql.plugin.jdbc.ConnectionFactory;
 import io.prestosql.plugin.jdbc.DriverConnectionFactory;
 import io.prestosql.plugin.jdbc.ForBaseJdbc;
 import io.prestosql.plugin.jdbc.JdbcClient;
+import io.prestosql.plugin.jdbc.TypeHandlingJdbcConfig;
 import io.prestosql.plugin.jdbc.credential.CredentialProvider;
 
 import static io.airlift.configuration.ConfigBinder.configBinder;
@@ -38,6 +39,7 @@ public class SqlServerClientModule
         binder.bind(Key.get(JdbcClient.class, ForBaseJdbc.class))
                 .to(SqlServerClient.class).in(Scopes.SINGLETON);
         configBinder(binder).bindConfig(BaseJdbcConfig.class);
+        configBinder(binder).bindConfig(TypeHandlingJdbcConfig.class);
     }
 
     @Provides
